Split eval script into logger and vizualizer
