+++ src/localization/scripts/data_log.py
@@ -1,13 +1,11 @@
 #!/usr/bin/env python
 
-import os
 import rospy
 import pandas as pd
-import matplotlib.pyplot as plt
 from nav_msgs.msg import Odometry
 
 
-class DataCollector():
+class DataLogger():
 
     def __init__(self):
         # Initialize ROS node
@@ -95,47 +93,6 @@ class DataCollector():
 
         rospy.loginfo("Successfully saved all data.")
 
-        DataVisualizer(
-            [self.gps_gt_path, self.gps_global_path, self.gps_local_path],
-            {
-                "names": ["ground_truth", "ekf_global", "ekf_local"]
-            }
-        )
-
-
-class DataVisualizer():
-
-    def __init__(self, file_paths, options=None):
-        # Save path of all data files
-        self.paths = file_paths
-
-        # Over-write provided options with defaults
-        # (see: https://www.stackabuse.com/how-to-merge-two-dictionaries-in-python/)
-        self.options = {**{
-            "names": ["ground_truth", "ekf_global", "ekf_local"],
-            "save_path": str(os.path.join(os.environ["AUTOMATED_HOME"], "data"))
-        }, **options}
-
-        # To store a list of DataFrames
-        self.data_frames = []
-
-        # Read data files
-        for file_path in self.paths:
-            df = pd.read_csv(file_path)
-            self.data_frames.append(df)
-
-        # Plot data
-        fig, axes = plt.subplots()
-        for i in range(len(self.data_frames)):
-            frame = self.data_frames[i]
-            frame.plot(x="x", y="y", ax=axes, xlabel="x", ylabel="y", label=options["names"][i])
-
-        # Save figure
-        fig_path = os.path.join(self.options["save_path"], "fig.png")
-        plt.savefig(fig_path)
-
-        print("Saved figure to: ", fig_path)
-
 
 if __name__ == "__main__":
-    DataCollector().run()
+    DataLogger().run()
